Logging: CliApplication refactor
Add logging in CliApplication
Move internal class CliParser out of CliApplication
Move ConfigResolving logic to ConfigResolvingLoaderProvider
